requirements.txt
fastapi==0.100.1
uvicorn==0.23.1
sqlalchemy==2.0.9
gunicorn==21.2.0
python-dotenv==1.0.0
alembic==1.11.1
asyncpg==0.28.0
email-validator==2.0.0.post2
passlib==1.7.4
PyJWT==2.8.0
httpx==0.24.1
anyio==3.7.1
minio==7.1.15
python-multipart==0.0.6
jinja2==3.1.2
fastapi-mail==1.4.1
pydantic==2.1.1
pydantic-settings==2.0.2
